Decreased the increments of noise

+++ src/simulation/simResult.h
@@ -72,7 +72,7 @@ struct SimResult {
     int totalSteps;
     bool allTargetsDestroyed;
     bool allSeekersDead;
-    int maxNoiseLevel;
+    double maxNoiseLevel;
 
     std::vector<SeekerResult> seekerResults;
     std::vector<TargetResult> targetResults;
